feat: only dhow disclaimer if SERVICE_FEE has been set

+++ lnbits/settings.py
@@ -1,8 +1,6 @@
 import importlib
 import os
 
-from decimal import Decimal
-
 
 wallets_module = importlib.import_module("lnbits.wallets")
 wallet_class = getattr(wallets_module, os.getenv("LNBITS_BACKEND_WALLET_CLASS", "LntxbotWallet"))
@@ -13,4 +11,4 @@ LNBITS_DATA_FOLDER = os.getenv("LNBITS_DATA_FOLDER", os.path.join(LNBITS_PATH, "
 WALLET = wallet_class()
 DEFAULT_WALLET_NAME = os.getenv("LNBITS_DEFAULT_WALLET_NAME", "LNbits wallet")
 FORCE_HTTPS = os.getenv("LNBITS_FORCE_HTTPS", "1") == "1"
-SERVICE_FEE = Decimal(os.getenv("LNBITS_SERVICE_FEE", "0.0"))
+SERVICE_FEE = float(os.getenv("LNBITS_SERVICE_FEE", "0.0"))
